Move factories and seeding to proper folder
